refactor(condo): DOMA-13015 refactored after review
