+++ apps/condo/domains/news/components/FilesPreview/ImageOrVideoPreview.module.css
@@ -1,0 +1,7 @@
+.image-or-video-preview {
+  width: 100%;
+  object-fit: cover;
+  border: 1px solid var(--condo-global-color-gray-3);
+  border-radius: 12px;
+  aspect-ratio: 16 / 9;
+}
