--- apps/condo/domains/news/components/FilesUploadList.module.css
@@ -1,134 +1,0 @@
-.upload-item {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 104px;
-  padding: 4px;
-  overflow: hidden;
-  background: #f5f5f5;
-  border-radius: 8px;
-  cursor: pointer;
-  transition: all 0.2s;
-}
-
-.upload-item--with-thumb {
-  height: 104px;
-}
-
-.upload-item--no-thumb {
-  height: 130px;
-}
-
-/* Контент */
-.media-wrapper {
-  position: relative;
-  width: 100%;
-  height: 104px;
-}
-
-.image {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  border-radius: 6px;
-}
-
-.icon-wrapper {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-  font-size: 48px;
-  background: #e0e0e0;
-  border-radius: 6px;
-}
-
-/* Название */
-.file-name {
-  width: 100%;
-  margin-top: 4px;
-  overflow: hidden;
-  font-size: 12px;
-  white-space: nowrap;
-  text-align: center;
-  text-overflow: ellipsis;
-}
-
-/* Длительность */
-.duration {
-  position: absolute;
-  right: 4px;
-  bottom: 4px;
-  padding: 2px 6px;
-  color: #fff;
-  font-size: 10px;
-  background: rgb(0 0 0 / 70%);
-  border-radius: 4px;
-}
-
-/* Overlay */
-.hover-overlay {
-  position: absolute;
-  display: flex;
-  gap: 8px;
-  align-items: center;
-  justify-content: center;
-  background-color: rgb(0 0 0 / 0%);
-  border-radius: 6px;
-  transition: background 0.2s;
-  inset: 0;
-}
-
-/* Кнопки */
-.action-button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 28px;
-  height: 28px;
-  font-size: 16px;
-  border-radius: 50%;
-  cursor: pointer;
-  opacity: 0;
-  transition: opacity 0.2s, transform 0.2s;
-}
-
-.eye-button {
-  color: #000;
-  background: rgb(255 255 255 / 85%);
-}
-
-.delete-button {
-  color: #fff;
-  background: rgb(255 0 0 / 85%);
-}
-
-/* Hover */
-.upload-item:hover .hover-overlay {
-  background-color: rgb(0 0 0 / 35%);
-}
-
-.upload-item:hover .action-button {
-  transform: scale(1.05);
-  opacity: 1;
-}
-
-.video-icon-wrapper {
-  position: relative;
-  width: 100%;
-  height: 100%;
-}
-
-.video-duration {
-  position: absolute;
-  right: 4px;
-  bottom: 4px;
-  padding: 2px 6px;
-  color: white;
-  font-size: 10px;
-  background: rgb(0 0 0 / 70%);
-  border-radius: 4px;
-}
