+++ apps/condo/domains/news/components/FilesUploadList/index.ts
@@ -1,0 +1,14 @@
+import { FilesUploadList } from './FilesUploadList'
+import { useModifiedFiles } from './hooks/useModifiedFiles'
+import { convertFilesToUploadType } from './utils/types'
+
+import type { DBFile, UploadFileType } from './FilesUploadList'
+import type { Action } from './hooks/useModifiedFiles'
+
+
+export { FilesUploadList }
+export { useModifiedFiles }
+export { convertFilesToUploadType }
+
+export type { DBFile, UploadFileType }
+export type { Action }
