+++ apps/condo/lang/en/en.json
@@ -1603,6 +1603,8 @@
   "news.create.title": "New news",
   "news.createNews": "Add news",
   "news.fields.files.label": "Add media files",
+  "news.fields.files.info.documents": "document {documentTypes} up to {documentSizeInMb} MB",
+  "news.fields.files.info.media": "Image {imageTypes} up to {imageSizeInMb} MB, video {videoTypes} up to {videoSizeInMb} MB",
   "news.fields.address.label": "Select addressees",
   "news.fields.body.error.length": "Please add message",
   "news.fields.body.placeholder": "News will appear in the 'News' section on the main screen, and a push notification will be sent to the resident",
@@ -3057,6 +3059,7 @@
   "pages.news.create.selectSharingApp.otherApps.description": "You can send news not only to mobile application, but also through other channels",
   "pages.news.newsItemCard.author": "author {createdBy}{isOwner, plural, other {} one { (you)}}",
   "pages.news.newsItemCard.field.body": "Text",
+  "pages.news.newsItemCard.field.files": "Media files",
   "pages.news.newsItemCard.field.sendAt": "Sent date",
   "pages.news.newsItemCard.field.title": "Title",
   "pages.news.newsItemCard.field.type": "Type",
