+++ apps/condo/lang/es/es.json
@@ -1603,6 +1603,8 @@
   "news.create.title": "Nueva noticia",
   "news.createNews": "Añadir noticia",
   "news.fields.files.label": "Agregar archivos multimedia",
+  "news.fields.files.info.documents": "documento {documentTypes} hasta {documentSizeInMb} MB",
+  "news.fields.files.info.media": "Imagen {imageTypes} hasta {imageSizeInMb} MB, vídeo {videoTypes} hasta {videoSizeInMb} MB",
   "news.fields.address.label": "Seleccionar destinatarios",
   "news.fields.body.error.length": "Por favor, añade el texto principal",
   "news.fields.body.placeholder": "El texto de la noticia aparecerá en la pantalla principal de la aplicación del vecino, en la sección ¨Noticias¨.\nTambién le enviaremos una notificación push.",
@@ -3055,6 +3057,7 @@
   "pages.news.create.selectSharingApp.otherApps.description": "Puedes enviar noticias no solo a la aplicación móvil, sino también a través de otros canales.\n",
   "pages.news.newsItemCard.author": "autor {createdBy}{isOwner, plural, other {} one { (tú)}}",
   "pages.news.newsItemCard.field.body": "Texto",
+  "pages.news.newsItemCard.field.files": "Archivos multimedia",
   "pages.news.newsItemCard.field.sendAt": "Fecha de envío",
   "pages.news.newsItemCard.field.title": "Título",
   "pages.news.newsItemCard.field.type": "Tipo",
