+++ apps/condo/lang/ru/ru.json
@@ -1603,6 +1603,8 @@
   "news.create.title": "Новая новость",
   "news.createNews": "Добавить новость",
   "news.fields.files.label": "Добавьте медиафайлы",
+  "news.fields.files.info.documents": "документ {documentTypes} до {documentSizeInMb} МБ",
+  "news.fields.files.info.media": "Изображение {imageTypes} до {imageSizeInMb} МБ, видео {videoTypes} до {videoSizeInMb} МБ",
   "news.fields.address.label": "Выберите адресатов",
   "news.fields.body.error.length": "Пожалуйста, добавьте основной текст",
   "news.fields.body.placeholder": "Текст новости появится на главном экране приложения жителя, в разделе «Новости».\nТакже отправим пользователю push-уведомление.",
@@ -3057,6 +3059,7 @@
   "pages.news.create.selectSharingApp.otherApps.description": "Отправлять новости можно не только в приложение Doma, но и по другим каналам",
   "pages.news.newsItemCard.author": "автор {createdBy}{isOwner, plural, other {} one { (вы)}}",
   "pages.news.newsItemCard.field.body": "Текст",
+  "pages.news.newsItemCard.field.files": "Медиафайлы",
   "pages.news.newsItemCard.field.sendAt": "Дата отправки",
   "pages.news.newsItemCard.field.title": "Заголовок",
   "pages.news.newsItemCard.field.type": "Тип",
